+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added support for sequence length warm-up.
 
+### Changed
+
+- Renamed `MemMapDataset` to `NumpyDataset`.
+
 ## [v1.0.6](https://github.com/allenai/OLMo-core/releases/tag/v1.0.6) - 2024-09-05
 
 ### Added